% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_stats}
\alias{add_stats}
\title{Create stats tables, adding calculated columns of counts (n) and comparison (match)}
\usage{
add_stats(ids)
}
\arguments{
\item{ids}{List of 2 tables, created using create_id()}
}
\value{
List of 2 tables, but with added columns (counts and matches)
}
\description{
Create stats tables, adding calculated columns of counts (n) and comparison (match)
}
\examples{
add_stats(ids)

}
