% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checker.R
\docType{class}
\name{Check_variable}
\alias{Check_variable}
\title{R6 Class helper for checking types and length of variables}
\arguments{
\item{object}{The object to check the type off.}

\item{type}{Single string. It's the type to check. It does not have to be the
exact string of the type to check, but it has to be close (ex: for integer
type, at least the string has to have "int" in it).
Basic types: char num int data.frame tibble list}

\item{len}{Integer. The length of the object to check.}

\item{len_operator}{Single string. It's the operator that we want to do the
length check.}

\item{compare_operator}{Unique string. It gives more options to compare the
.len field of our class.}
}
\value{
A checker object

Boolean. Both type and length need to be true.
@description
It will compare the type of the field object, with the fild type.
Then update the check_type field.
@return Boolean

Boolean

tibble
}
\description{
Create object of the class. It needs to have at least a type or len arguments.

Prints get_values, and the constructor inputs.

It will coompare length, type or both fields, if provided on the constructor.

It will compare the length of the field object, and use
the compare_operator to do the comparison. Then update the check_len field.

Show the 3 results of the comparison.

Show the global results of the comparison

Show the type results of the comparison

Show the length results of the comparison
}
\section{Fields}{

\describe{
\item{\code{.object}}{The object that we want to check}

\item{\code{.type}}{The type to check. String with the name of the class.
To do the check, it needs to match at least a part of a string
(ex: for integer type, at least the string has to have "int" in it).}

\item{\code{len}}{The length of the object to check.}

\item{\code{.check_all}}{Result of a global check, it has to match what it checked (len, or type, or both).}

\item{\code{.check_type}}{Result of a type check.}

\item{\code{.check_len}}{Result of a length check.}
}}

