% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checker.R
\name{checkvar}
\alias{checkvar}
\title{Check the type of a basic variable.}
\usage{
checkvar(object, type = NULL, len = NULL, len_operator = "==")
}
\arguments{
\item{object}{The object to check the type off.}

\item{type}{Single string. It's the type to check. It does not have to be the
exact string of the type to check, but it has to be close (ex: for integer
type, at least the string has to have "int" in it).
Basic types: char num int data.frame tibble list}

\item{len}{Integer. The length of the object to check.}

\item{len_operator}{Single string. It's the operator that we want to do the
length check.}
}
\value{
Boolean
}
\description{
It is used internaly on the package, for consistency. It's a wrap up of the
\code{Check_variable} class.
}
\examples{
# Check the type
checkvar(123, type = "integer")

# The type can be any string, if it contains "int"
checkvar(123, type = "int")

# Check the length
checkvar(1:2, len = 2)

# We can change the operator to check
checkvar(1:2, len = 5, len_operator = "<")

# We could check everything, and the type and length needs to match
checkvar(1:2, type = "int", len = 2)

}
