% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{create_id}
\alias{create_id}
\title{Create an id for both tables, pasting columns together}
\usage{
create_id(table, col_pos, sep = " ")
}
\arguments{
\item{sep}{}
}
\description{
Create an id for both tables, pasting columns together
}
