% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_duplicate}
\alias{is_duplicate}
\title{Check if a "CompareID" object has duplicates}
\usage{
is_duplicate(objecte, view = FALSE)
}
\arguments{
\item{objecte}{CompareID}

\item{view}{Boolean. If we want to view witch values are duplicates}
}
\value{
Boolean vector. Tells if id1 or id2 has duplicates
}
\description{
Check if a "CompareID" object has duplicates
}
\examples{
is_duplicate(CompareID$new(table1, table2, "municipi"))

}
