% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_id.R
\name{length_id}
\alias{length_id}
\title{Utils check for select id}
\usage{
length_id(id)
}
\arguments{
\item{id}{The column id to select. Must be a char, factor, num or int
of length 1 or 2 (from the two tables).}
}
\value{
Length of the id (1, 2 or error)
}
\description{
Utils check for select id
}
