% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(
  path,
  format,
  setclass = getOption("rio.import.class", "data.frame"),
  which,
  ...
)
}
\arguments{
\item{path}{A character vector with one element. Path to the file.}

\item{format}{Read rio::import()}

\item{setclass}{Read rio::import()}

\item{which}{Read rio::import()}

\item{...}{Read rio::import()}
}
\value{
A data.frame
}
\description{
Load data
}
\examples{
path = paste0("https://www.stats.govt.nz/assets/Uploads/Annual-enterprise-survey",
              "/Annual-enterprise-survey-2021-financial-year-provisional",
              "/Download-data/annual-enterprise-survey-2021-financial-year-provisional-csv.csv")
load_data(path)
}
