% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_id.R
\name{select_id}
\alias{select_id}
\title{Select which column pick as id}
\usage{
select_id(table1, table2, id)
}
\arguments{
\item{table1}{A data.frame}

\item{table2}{A data.frame}

\item{id}{The column id to select. Must be a vector of length 1 or 2}
}
\value{
A list with the 2 columns, in vector format
}
\description{
Select which column pick as id
}
