% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step_manager.R
\docType{class}
\name{Step_manager}
\alias{Step_manager}
\title{R6 Class Step_manager, will help manage the steps: add, remove, sort...}
\arguments{
\item{row_number}{Int, of length one or more}

\item{from}{int, row that we want to move}

\item{to}{int, destination of the row that we want to move}

\item{dades}{tibble. It needs a column with name "id".}
}
\value{
A Step_manager object.

An integer.

A vector with the info
}
\description{
Create object of the class.

Adds a step at the end of the process

Remove one or more steps, based on the index of the table.

Move a step, from one row to another.

Execute the steps in order.

Print the info method

Displays the number of steps of the process.

Show basic info
}
\section{Fields}{

\describe{
\item{\code{.steps}}{Tibble. A table containing the steps,
and the order which they are applied}
}}

