% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table1}
\alias{table1}
\title{Basic testing table1}
\format{
\subsection{\code{table1}}{

A data frame with 9 rows and 2 columns:

\describe{
\item{municipi}{The city name, or the id that we want to match}
\item{info}{Some random info, in this case the row position}
}
}
}
\usage{
table1
}
\description{
A small table, used on test and the Basic vignette.
It will be used on the left part of the left join
}
\keyword{datasets}
