% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table2}
\alias{table2}
\title{Basic testing table2}
\format{
\subsection{\code{table1}}{

A data frame with 13 rows and 2 columns:

\describe{
\item{municipi}{The city name, or the id that we want to match}
\item{info}{Some random info, in this case the row position}
}
}
}
\usage{
table2
}
\description{
A small table, used on test and the Basic vignette.
It will be used on the right part of the left join. It has duplicate values.
}
\keyword{datasets}
