% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_complex1}
\alias{table_complex1}
\title{Complex example, table_complex1}
\format{
\subsection{\code{table_complex1}}{

A data frame with 928 rows and 2 columns:

\describe{
\item{municipi}{The city name, or the id that we want to match}
\item{info}{Some random info, in this case a number}
}
}
}
\usage{
table_complex1
}
\description{
A table with some municipalities of Catalonia
It will be used on the left part of the left join.
}
\keyword{datasets}
