% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_complex2}
\alias{table_complex2}
\title{Complex example, table_complex2}
\format{
\subsection{\code{table_complex2}}{

A data frame with 956 rows and 5 columns:

\describe{
\item{municipi}{The municipality name, or the id that we want to match}
\item{provincia}{The province of the municipality}
\item{comarques}{Region name}
\item{areaPolicial}{Smaller police area}
\item{regioPolicial}{Larger police area}
}
}
}
\source{
\url{https://www.idescat.cat/codis/?id=50&n=9}
}
\usage{
table_complex2
}
\description{
A table with all the municipalities of Catalonia, and with some more
geografic information
It will be used on the right part of the left join, so we can add information
to the base dataset. It has some duplicates.
}
\keyword{datasets}
