% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareID.R
\docType{class}
\name{CompareID}
\alias{CompareID}
\title{R6 Class. Principal class, to preprocess the tables.}
\arguments{
\item{dades1}{data.frame. Taula en la cual volem extreure el primer identificador}

\item{dades2}{data.frame. Taula en la cual volem extreure el segon identificador}
}
\value{
A CompareID object

Updates self$join_result, and returns invisible(self).

data.frame

data.frame
}
\description{
Create object of the class.

Do a join with id1 and id2. The type of join is chose by join_type.
Can add more parameters (see ?dplyr::left_join)
TODO: Add the preprocessing step.

Return the missing matches of the join

Return the matches of the join
}
\section{Fields}{

\describe{
\item{\code{id1}}{Column identifier of the first table}

\item{\code{id2}}{Column identifier of the second table}

\item{\code{id_name}}{Character. The name of the column identifiers. Can be
a vector of length 1 or 2, depending if the name is the same or not.}

\item{\code{join-type}}{String. Type of join to use (left, right, full, inner)}
}}

\examples{
CompareID$new(table1, table2, "municipi")
CompareID$new(table1, table2, "municipi")$join()$join_result
CompareID$new(table1, table2, "municipi")$join()$join_miss
CompareID$new(table1, table2, "municipi")$join()$join_miss
}
