% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_duplicate}
\alias{is_duplicate}
\title{Check if a tibble has duplicates}
\usage{
is_duplicate(taula)
}
\arguments{
\item{taula}{Tibble with 1 column, or multiple columns with one named "n"}
}
\value{
Boolean: T if duplicates, F if unique.
}
\description{
Check if a tibble has duplicates
}
\examples{
is_duplicate(table2, "municipi")

}
