% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step.R
\docType{class}
\name{Step}
\alias{Step}
\title{R6 Class Step, will help adding a step on the preprocessing process}
\arguments{
\item{type}{string. Select the type of step/function to apply.
Can be: c("remove")}

\item{...}{Other arguments of the function that will be applied.
\itemize{
\item remove: pattern, position
\item 
}}

\item{pattern}{string. Regex to apply}

\item{position}{string. Where to apply the pattern.
Can be: c("all", "one", "start", "end")}
}
\value{
A Step object.

function
}
\description{
Create object of the class.

Will help choosing the function step to apply (equivalence table from self$.type)

If type = remove, will remove a pattern from a dataset dataset.
}
\section{Fields}{

\describe{
\item{\code{type}}{char. The type of step to apply.
Can be: c("remove")}

\item{\code{.step}}{function. A step to aply to CompareID}

\item{\code{.pattern}}{char. The pattern to apply to some steps}

\item{\code{.position}}{char. The position to apply the pattern.
Can be: c("all", "one", "start", "end")}
}}

\examples{
Step$new(type = "remove", pattern = "a")
}
